Move fonduer_model.py to fonduer.packaging
