+++ tests/shared/hardware_fonduer_model.py
@@ -6,8 +6,8 @@ from pandas import DataFrame
 
 import tests.shared.hardware_subclasses  # noqa
 from fonduer.learning.dataset import FonduerDataset
+from fonduer.packaging.fonduer_model import F_matrix, FonduerModel
 from fonduer.parser.models import Document
-from fonduer.utils.fonduer_model import F_matrix, FonduerModel
 from tests.shared.hardware_lfs import TRUE
 from tests.shared.hardware_utils import get_implied_parts
